+++ server/src/main/java/com/rezero/inandout/expense/service/base/impl/ExpenseServiceImpl.java
@@ -1,4 +1,4 @@
-package com.rezero.inandout.expense.service.impl;
+package com.rezero.inandout.expense.service.base.impl;
 
 import com.rezero.inandout.exception.ExpenseException;
 import com.rezero.inandout.exception.errorcode.ExpenseErrorCode;
@@ -9,7 +9,7 @@ import com.rezero.inandout.expense.model.*;
 import com.rezero.inandout.expense.repository.DetailExpenseCategoryRepository;
 import com.rezero.inandout.expense.repository.ExpenseCategoryRepository;
 import com.rezero.inandout.expense.repository.ExpenseRepository;
-import com.rezero.inandout.expense.service.ExpenseService;
+import com.rezero.inandout.expense.service.base.ExpenseService;
 import com.rezero.inandout.member.entity.Member;
 import com.rezero.inandout.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
