+++ server/src/main/java/com/rezero/inandout/expense/service/report/ExpenseReportService.java
@@ -1,0 +1,4 @@
+package com.rezero.inandout.expense.service.report;
+
+public interface ExpenseReportService {
+}
