+++ server/src/main/java/com/rezero/inandout/expense/service/report/impl/ExpenseReportServiceImpl.java
@@ -1,0 +1,4 @@
+package com.rezero.inandout.expense.service.report.impl;
+
+public class ExpenseReportServiceImpl {
+}
