+++ server/src/main/java/com/rezero/inandout/expense/service/table/ExpenseTableService.java
@@ -1,4 +1,4 @@
-package com.rezero.inandout.expense.service;
+package com.rezero.inandout.expense.service.table;
 
 import com.rezero.inandout.expense.model.CategoryAndExpenseDto;
 import com.rezero.inandout.expense.model.ExpenseInput;
@@ -6,7 +6,7 @@ import com.rezero.inandout.expense.model.ExpenseInput;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ExpenseGraphService {
+public interface ExpenseTableService {
     void addAndUpdateExpense(String email, List<ExpenseInput> inputs);
 
     CategoryAndExpenseDto getCategoryAndExpenseDto(String email, LocalDate startDt, LocalDate endDt);
