+++ server/src/main/java/com/rezero/inandout/income/service/report/IncomeReportService.java
@@ -1,0 +1,4 @@
+package com.rezero.inandout.income.service.report;
+
+public interface IncomeReportService {
+}
