+++ server/src/main/java/com/rezero/inandout/income/service/report/impl/IncomeReportServiceImpl.java
@@ -1,0 +1,4 @@
+package com.rezero.inandout.income.service.report.impl;
+
+public class IncomeReportServiceImpl {
+}
