+++ server/src/main/java/com/rezero/inandout/income/service/table/IncomeTableService.java
@@ -1,4 +1,4 @@
-package com.rezero.inandout.income.service;
+package com.rezero.inandout.income.service.table;
 
 
 import com.rezero.inandout.income.model.CategoryAndIncomeDto;
@@ -7,7 +7,7 @@ import com.rezero.inandout.income.model.IncomeInput;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface IncomeGraphService {
+public interface IncomeTableService {
 
 
     void addAndUpdateIncome(String email, List<IncomeInput> incomeInputList);
