+++ server/src/test/java/com/rezero/inandout/expense/controller/ExpenseControllerTest.java
@@ -2,8 +2,8 @@ package com.rezero.inandout.expense.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.rezero.inandout.expense.model.*;
-import com.rezero.inandout.expense.service.ExpenseGraphService;
-import com.rezero.inandout.expense.service.ExpenseService;
+import com.rezero.inandout.expense.service.table.ExpenseTableService;
+import com.rezero.inandout.expense.service.base.ExpenseService;
 import com.rezero.inandout.member.entity.Member;
 import com.rezero.inandout.member.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
@@ -42,7 +42,7 @@ class ExpenseControllerTest {
     @MockBean
     private ExpenseService expenseService;
     @MockBean
-    private ExpenseGraphService expenseGraphService;
+    private ExpenseTableService expenseTableService;
 
     @MockBean
     private MemberService memberService;
@@ -98,7 +98,7 @@ class ExpenseControllerTest {
         ArgumentCaptor<List<ExpenseInput>> captor = ArgumentCaptor.forClass(List.class);
 
         //then
-        verify(expenseGraphService, times(1)).addAndUpdateExpense(any(), captor.capture());
+        verify(expenseTableService, times(1)).addAndUpdateExpense(any(), captor.capture());
         assertEquals(captor.getValue().get(0).getExpenseItem(), "쌀과자");
         assertEquals(captor.getValue().get(1).getExpenseItem(), "만두");
     }
@@ -153,7 +153,7 @@ class ExpenseControllerTest {
                 .expenseDtos(expenseDtos)
                 .build();
 
-        given(expenseGraphService.getCategoryAndExpenseDto(anyString(), any(), any()))
+        given(expenseTableService.getCategoryAndExpenseDto(anyString(), any(), any()))
             .willReturn(categoryAndExpenseDto);
 
         //when
