+++ server/src/test/java/com/rezero/inandout/expense/service/impl/ExpenseServiceImplTest.java
@@ -11,6 +11,7 @@ import com.rezero.inandout.expense.model.ExpenseInput;
 import com.rezero.inandout.expense.repository.DetailExpenseCategoryRepository;
 import com.rezero.inandout.expense.repository.ExpenseCategoryRepository;
 import com.rezero.inandout.expense.repository.ExpenseRepository;
+import com.rezero.inandout.expense.service.base.impl.ExpenseServiceImpl;
 import com.rezero.inandout.member.entity.Member;
 import com.rezero.inandout.member.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
